move get_library to _ffi.py
