+++ oscrypto/_openssl/_libcrypto_cffi.py
@@ -4,8 +4,8 @@ from __future__ import unicode_literals, division, absolute_import, print_functi
 import re
 
 from .. import _backend_config
-from ..util import get_library
 from .._errors import pretty_message
+from .._ffi import get_library
 from .._ffi import register_ffi
 from ..errors import LibraryNotFoundError
