+++ oscrypto/util.py
@@ -2,9 +2,6 @@
 from __future__ import unicode_literals, division, absolute_import, print_function
 
 import sys
-import platform
-
-from ctypes.util import find_library
 
 from ._errors import pretty_message
 from ._types import type_name, byte_cls
@@ -20,7 +17,6 @@ else:
 __all__ = [
     'constant_compare',
     'rand_bytes',
-    'get_library',
 ]
 
 
@@ -65,28 +61,3 @@ def constant_compare(a, b):
     for x, y in zip(a, b):
         result |= x ^ y
     return result == 0
-
-
-def get_library(name, unversioned, fallback):
-    """
-    Retrieves the C library, falling back to a specified library if one is not found
-
-    :param name:
-        The library to search the system for
-
-    :param unversioned:
-        The unversioned library we don't want to use
-
-    :param fallback:
-        Fallback library when we don't find a suitable library to use
-
-    :return:
-        Path to the library
-    """
-    library = find_library(name)
-    if not library and sys.platform == 'darwin' and tuple(map(int,  platform.mac_ver()[0].split('.'))) >= (10, 16):
-        library == fallback
-    elif sys.platform == 'darwin' and tuple(map(int,  platform.mac_ver()[0].split('.'))) >= (10, 15) and \
-            library == unversioned:
-        library = fallback
-    return library
